requirements.txt
Django==3.1.0
djongo==1.3.3
sqlparse==0.2.4
djangorestframework==3.11.1
psycopg2-binary==2.8.4
otpauth==1.0.1
Pillow==7.2.0
numpy==1.19.2
requests==2.24.0
cryptography==3.0
tensorflow==2.3.0
scikit-image==0.17.2
opencv-python==4.4.0.46
imutils==0.5.3
opencv-contrib-python
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
pika==1.1.0

